+++ src/cross_references.jl
@@ -498,13 +498,13 @@ function find_object(doc::Documenter.Document, binding, typesig)
                 # We've found an actual match out of the possible choices! Use it.
                 return candidate
             else
-                # Special case: UnionAll signature might be equivalent to an existing object, but not identical
-                if isa(typesig, UnionAll) 
+                # Special case: UnionAll signature might be equivalent to an existing object, but not identical ([#2889])
+                if isa(typesig, UnionAll)
                     for object in objects
                         if object.binding === candidate.binding &&
-                           object.signature == candidate.signature && # equality, not identity!
-                           object.noncanonical_extra === candidate.noncanonical_extra
-                           return object
+                                object.signature == candidate.signature && # equality, not identity!
+                                object.noncanonical_extra === candidate.noncanonical_extra
+                            return object
                         end
                     end
                 end